% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_data.R
\docType{data}
\name{hobo_rmbl_data}
\alias{hobo_rmbl_data}
\title{HOBO RMBL data.
Model files for measure the snowpack duration at RMBL plots. Use
these examples to work with the \code{snowpack} package.}
\format{
A \code{dataframe tibble}

An object of class \code{list} of length 2.
}
\usage{
hobo_rmbl_data
}
\description{
HOBO RMBL data.
Model files for measure the snowpack duration at RMBL plots. Use
these examples to work with the \code{snowpack} package.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list with two-element: road and pbm, cleaned files  from HOBO data loggers.}
}}

\keyword{datasets}
