% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_full_range.R
\name{hobo_summary}
\alias{hobo_summary}
\title{HOBO summary}
\usage{
hobo_summary(df)
}
\arguments{
\item{df}{A data.frame formatted as the example dataset. Review
hobo_rmbl_data to see details.}
}
\value{
a tibble \code{HOBO} summary object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function builds a summary of row values. By default
temperature measured in Fahrenheit degrees, it's omitted.
}
\examples{
data("hobo_rmbl_data")
road <- hobo_rmbl_data[[1]]
road \%>\%
hobo_summary()
}
