% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hobo_rmbl.R
\name{read_hobo_rmbl}
\alias{read_hobo_rmbl}
\title{Read HOBO data}
\usage{
read_hobo_rmbl(path_file, write = FALSE, ...)
}
\arguments{
\item{path_file}{}

\item{write}{}

\item{...}{}
}
\value{

}
\description{
Read HOBO data
}
\examples{
Write clean files
}
