% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_pack_grouped.R
\name{snow_pack_grouped}
\alias{snow_pack_grouped}
\title{Grouped snow pack}
\usage{
snow_pack_grouped(df, variable, value)
}
\arguments{
\item{df}{\verb{HOBO summary object}}

\item{variable}{The options for this parameter are:
\code{mean_temp} \code{min_temp} \code{max_temp} \code{range}}

\item{value}{this parameter defines the range of temperature variations}
}
\value{
a tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
this function takes a subset of observations  between a temperature interval
}
\examples{
data("hobo_rmbl_data")
road <- hobo_rmbl_data[[1]]
road \%>\%
 hobo_summary() \%>\%
 snow_pack_grouped(mean_temp, 2)
}
