% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_snow_pack_grouped.R
\name{summary_snow_pack_grouped}
\alias{summary_snow_pack_grouped}
\title{Summarise snow pack grouped}
\usage{
summary_snow_pack_grouped(df, by = hobolocation, date = longdate)
}
\arguments{
\item{df}{\code{snow_pack_grouped} function output}

\item{by}{sensor position}

\item{date}{date variable column}
}
\value{
A summary tibble with the start and end date of snow
presence, and the number of days.
}
\description{
Summarise snow pack grouped
}
\examples{
data("hobo_rmbl_data")
pbm <- hobo_rmbl_data[[2]]

pbm \%>\%
hobo_summary() \%>\%
snow_pack_grouped(mean_temp, 2) \%>\%
summary_snow_pack_grouped()

}
