% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_snow_pack.R
\name{summary_snow_pack}
\alias{summary_snow_pack}
\title{Summarise snow pack}
\usage{
summary_snow_pack(df, location = hobolocation, date = longdate)
}
\arguments{
\item{date}{}
}
\value{

}
\description{
Summarise snow pack
}
